103-keygen: match crackme5 with signed int arithmetic

+++ doubly_linked_lists/103-keygen.c
@@ -12,7 +12,7 @@ int main(int ac, char **av)
 {
 	char key[7];
 	const char *u;
-	unsigned int len, i, sum, prod, maxc, sumsqr, r;
+	int len, i, sum, prod, maxc, sumsqr, r;
 	const char *alph =
 		"A-CHRDw87lNS0E9B2TibgpnMVys5XzvtOGJcYLU+4mjW6fxqZ";
 
@@ -23,45 +23,45 @@ int main(int ac, char **av)
 	}
 
 	u = av[1];
-	len = (unsigned int)strlen(u);
+	len = (int)strlen(u);
 
-	/* k0: length */
-	key[0] = alph[(len ^ 59U) & 63U];
+	/* k0: from length */
+	key[0] = alph[(len ^ 59) & 63];
 
 	/* k1: sum of chars */
 	sum = 0;
 	for (i = 0; i < len; i++)
-		sum += (unsigned int)u[i];
-	key[1] = alph[(sum ^ 79U) & 63U];
+		sum += (int)u[i];
+	key[1] = alph[(sum ^ 79) & 63];
 
-	/* k2: product of chars (overflow intentional) */
+	/* k2: product of chars (signed overflow intentional to match crackme) */
 	prod = 1;
 	for (i = 0; i < len; i++)
-		prod *= (unsigned int)u[i];
-	key[2] = alph[(prod ^ 85U) & 63U];
+		prod *= (int)u[i];
+	key[2] = alph[(prod ^ 85) & 63];
 
 	/* k3: rng with (max char ^ 14) */
 	maxc = 0;
 	for (i = 0; i < len; i++)
-		if ((unsigned int)u[i] > maxc)
-			maxc = (unsigned int)u[i];
-	srand(maxc ^ 14U);
-	key[3] = alph[rand() & 63U];
+		if ((int)u[i] > maxc)
+			maxc = (int)u[i];
+	srand(maxc ^ 14);
+	key[3] = alph[rand() & 63];
 
 	/* k4: sum of squares */
 	sumsqr = 0;
 	for (i = 0; i < len; i++)
 	{
-		unsigned int c = (unsigned int)u[i];
+		int c = (int)u[i];
 
 		sumsqr += c * c;
 	}
-	key[4] = alph[(sumsqr ^ 239U) & 63U];
+	key[4] = alph[(sumsqr ^ 239) & 63];
 
-	/* k5: advance rng username[0] times */
-	for (i = 0, r = 0; i < (unsigned int)u[0]; i++)
-		r = (unsigned int)rand();
-	key[5] = alph[(r ^ 229U) & 63U];
+	/* k5: advance rng u[0] times */
+	for (i = 0, r = 0; i < (int)u[0]; i++)
+		r = rand();
+	key[5] = alph[(r ^ 229) & 63];
 
 	key[6] = '\0';
 	printf("%s", key);
